chore: rename __test__ to test; chore: delete old test;
