+++ components/Airdrop/index.test.js
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import Airdrop from '.';
-import { dummyAccount, dummyNetwork } from '../../__test__/utils';
+import { dummyAccount, dummyNetwork } from '../../test/utils';
 
 jest.mock('@solana/web3.js', () => ({
   ...jest.requireActual('@solana/web3.js'),
